+++ src/main/java/com/study/studyproject/member/service/MemberService.java
@@ -1,9 +1,11 @@
 package com.study.studyproject.member.service;
 
 
+import com.study.studyproject.global.oauth.OAuthAttributes;
 import com.study.studyproject.member.domain.Member;
 import com.study.studyproject.list.dto.ListResponseDto;
 import com.study.studyproject.global.GlobalResultDto;
+import com.study.studyproject.member.domain.SocialType;
 import com.study.studyproject.member.dto.MemberListRequestDto;
 import com.study.studyproject.member.dto.MemberUpdateResDto;
 import com.study.studyproject.member.dto.UserInfoResponseDto;
@@ -21,9 +23,9 @@ public interface MemberService {
     GlobalResultDto userInfoUpdate(Member member , MemberUpdateResDto memberUpdateResDto);
 
     Page<ListResponseDto> listMember(Long token, MemberListRequestDto memberListRequestDto, Pageable pageable);
-    //
 
 
+    Member getOrCreateUser(OAuthAttributes attributes, SocialType socialType);
 
 
 }
